Apo/requirements.txt
google-generativeai>=0.3.0
opencv-python>=4.8.0
numpy>=1.24.0
fer>=22.5.0 